Edit image to correct alignment issues
Resize/replace strawberry-ice-cream.jpg to fix width-height ratio
issues. Previously, since it was a larger size than the other ice cream
images, the slight difference in its height and width caused it to
appear unaligned on the webpage with the other images. Making the image
smaller, and equating its width/height, corrects this.